chore: move renderers to a dedicated folder
