+++ cli.ts
@@ -7,17 +7,17 @@ import minimist from 'minimist'
 import { red, yellow, bold, green, cyan } from 'kolorist'
 
 import { getCommand } from './lib/get-command'
-import { renderApp } from './lib/render-app'
-import { renderLib } from './lib/render-lib'
-import { renderMonorepo } from './lib/render-monorepo'
-import { renderReadme } from './lib/render-readme'
+import { renderApp } from './lib/renderers/app'
+import { renderLib } from './lib/renderers/lib'
+import { renderMonorepo } from './lib/renderers/monorepo'
+import { renderReadme } from './lib/renderers/readme'
+import { renderWithHusky } from './lib/renderers/wtih-husky'
+import { renderWithActions } from './lib/renderers/wtih-actions'
+import { renderWithChangesets } from './lib/renderers/wtih-changesets'
 import { gitInitRepo } from './lib/git-init-repo'
 import { getUserAgent } from './lib/get-user-agent'
-import { renderWithHusky } from './lib/render-wtih-husky'
 import { canSkipEmptying } from './lib/can-skip-emptying'
 import { emptyDir } from './lib/empty-dir'
-import { renderWithActions } from './lib/render-wtih-actions'
-import { renderWithChangesets } from './lib/render-wtih-changesets'
 
 function isValidPackageName(projectName: string) {
   return /^(?:@[a-z0-9-*~][a-z0-9-*._~]*\/)?[a-z0-9-~][a-z0-9-._~]*$/.test(
