+++ lib/renderers/package-json.ts
